Removes plugin functionlity and moves MavenResolver and database connectors to core
