+++ analyzer/metadata-plugin/src/main/java/eu/fasten/analyzer/metadataplugin/Main.java
@@ -18,7 +18,7 @@
 
 package eu.fasten.analyzer.metadataplugin;
 
-import eu.fasten.server.connectors.PostgresConnector;
+import eu.fasten.core.dbconnectors.PostgresConnector;
 import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.sql.SQLException;
