+++ analyzer/pom-analyzer/src/main/java/eu/fasten/analyzer/pomanalyzer/Main.java
@@ -18,7 +18,7 @@
 
 package eu.fasten.analyzer.pomanalyzer;
 
-import eu.fasten.server.connectors.PostgresConnector;
+import eu.fasten.core.dbconnectors.PostgresConnector;
 import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.sql.SQLException;
