+++ analyzer/pom.xml
@@ -22,7 +22,6 @@
         <module>graph-plugin</module>
         <module>pom-analyzer</module>
         <module>repo-cloner-plugin</module>
-        <module>maven-resolver</module>
         <module>restapi-plugin</module>
     </modules>
