+++ analyzer/restapi-plugin/src/main/java/eu/fasten/analyzer/restapiplugin/Main.java
@@ -18,7 +18,7 @@
 
 package eu.fasten.analyzer.restapiplugin;
 
-import eu.fasten.server.connectors.PostgresConnector;
+import eu.fasten.core.dbconnectors.PostgresConnector;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import picocli.CommandLine;
